+++ e3nn/nn/_batchnorm.py
@@ -28,20 +28,28 @@ class BatchNorm(nn.Module):
 
     reduce : {'mean', 'max'}
         method used to reduce
+
+    instance : bool
+        apply instance norm instead of batch norm
     """
-    def __init__(self, irreps, eps=1e-5, momentum=0.1, affine=True, reduce='mean', normalization='component'):
+    def __init__(self, irreps, eps=1e-5, momentum=0.1, affine=True, reduce='mean', instance=False, normalization='component'):
         super().__init__()
 
         self.irreps = o3.Irreps(irreps)
         self.eps = eps
         self.momentum = momentum
         self.affine = affine
+        self.instance = instance
 
-        num_scalar = sum(mul for mul, ir in self.irreps if ir.l == 0)
+        num_scalar = sum(mul for mul, ir in self.irreps if ir.is_scalar())
         num_features = self.irreps.num_irreps
 
-        self.register_buffer('running_mean', torch.zeros(num_scalar))
-        self.register_buffer('running_var', torch.ones(num_features))
+        if self.instance:
+            self.register_buffer('running_mean', None)
+            self.register_buffer('running_var', None)
+        else:
+            self.register_buffer('running_mean', torch.zeros(num_scalar))
+            self.register_buffer('running_var', torch.ones(num_features))
 
         if affine:
             self.weight = nn.Parameter(torch.ones(num_features))
@@ -79,12 +87,9 @@ class BatchNorm(nn.Module):
         batch, *size, dim = input.shape
         input = input.reshape(batch, -1, dim)  # [batch, sample, stacked features]
 
-        if self.training:
+        if self.training and not self.instance:
             new_means = []
             new_vars = []
-        else:
-            new_means = None
-            new_vars = None
 
         fields = []
         ix = 0
@@ -101,20 +106,23 @@ class BatchNorm(nn.Module):
             # [batch, sample, mul, repr]
             field = field.reshape(batch, -1, mul, d)
 
-            if ir.l == 0:  # scalars
-                if self.training:
-                    field_mean = field.mean([0, 1]).reshape(mul)  # [mul]
-                    new_means.append(
-                        self._roll_avg(self.running_mean[irm:irm + mul], field_mean)
-                    )
+            if ir.is_scalar():  # scalars
+                if self.training or self.instance:
+                    if self.instance:
+                        field_mean = field.mean(1).reshape(batch, mul)  # [batch, mul]
+                    else:
+                        field_mean = field.mean([0, 1]).reshape(mul)  # [mul]
+                        new_means.append(
+                            self._roll_avg(self.running_mean[irm:irm + mul], field_mean)
+                        )
                 else:
                     field_mean = self.running_mean[irm: irm + mul]
                 irm += mul
 
                 # [batch, sample, mul, repr]
-                field = field - field_mean.reshape(mul, 1)
+                field = field - field_mean.reshape(-1, 1, mul, 1)
 
-            if self.training:
+            if self.training or self.instance:
                 if self.normalization == 'norm':
                     field_norm = field.pow(2).sum(3)  # [batch, sample, mul]
                 elif self.normalization == 'component':
@@ -129,23 +137,24 @@ class BatchNorm(nn.Module):
                 else:
                     raise ValueError("Invalid reduce option {}".format(self.reduce))
 
-                field_norm = field_norm.mean(0)  # [mul]
-                new_vars.append(self._roll_avg(self.running_var[irv: irv + mul], field_norm))
+                if not self.instance:
+                    field_norm = field_norm.mean(0)  # [mul]
+                    new_vars.append(self._roll_avg(self.running_var[irv: irv + mul], field_norm))
             else:
                 field_norm = self.running_var[irv: irv + mul]
             irv += mul
 
-            field_norm = (field_norm + self.eps).pow(-0.5)  # [mul]
+            field_norm = (field_norm + self.eps).pow(-0.5)  # [(batch,) mul]
 
             if self.affine:
                 weight = self.weight[iw: iw + mul]  # [mul]
                 iw += mul
 
-                field_norm = field_norm * weight  # [mul]
+                field_norm = field_norm * weight  # [(batch,) mul]
 
-            field = field * field_norm.reshape(mul, 1)  # [batch, sample, mul, repr]
+            field = field * field_norm.reshape(-1, 1, mul, 1)  # [batch, sample, mul, repr]
 
-            if self.affine and d == 1:  # scalars
+            if self.affine and ir.is_scalar():  # scalars
                 bias = self.bias[ib: ib + mul]  # [mul]
                 ib += mul
                 field += bias.reshape(mul, 1)  # [batch, sample, mul, repr]
@@ -157,14 +166,14 @@ class BatchNorm(nn.Module):
             msg = fmt.format(dim, ix)
             raise AssertionError(msg)
 
-        if self.training:
+        if self.training and not self.instance:
             assert irm == self.running_mean.numel()
             assert irv == self.running_var.size(0)
         if self.affine:
             assert iw == self.weight.size(0)
             assert ib == self.bias.numel()
 
-        if self.training:
+        if self.training and not self.instance:
             torch.cat(new_means, out=self.running_mean)
             torch.cat(new_vars, out=self.running_var)