+++ e3nn/o3/_reduce.py
@@ -86,15 +86,15 @@ class ReducedTensorProducts(fx.GraphModule):
         For instance ``ij=ji`` means that the tensor has to indices and if they are exchanged, its value is the same.
         ``ij=-ji`` means that the tensor change its sign if the two indices are exchanged.
 
-    irreps : dict of `Irreps`
-        each letter present in the formula has to be present in the ``irreps`` dictionary, unless it can be inferred by the formula.
-        For instance if the formula is ``ij=ji`` you can provide the representation of ``i`` only: ``irreps = {'i': o3.Irreps(...)}``.
-
     filter_ir_out : list of `Irrep`, optional
         Optional, list of allowed irrep in the output
 
     filter_ir_mid : list of `Irrep`, optional
         Optional, list of allowed irrep in the intermediary operations
+
+    **kwargs : dict of `Irreps`
+        each letter present in the formula has to be present in the ``irreps`` dictionary, unless it can be inferred by the formula.
+        For instance if the formula is ``ij=ji`` you can provide the representation of ``i`` only: ``ReducedTensorProducts('ij=ji', i='1o')``.
 
     Attributes
     ----------
@@ -262,9 +262,14 @@ class ReducedTensorProducts(fx.GraphModule):
         outs = []
         for vp_list in outputs:
             v, p = vp_list[0]
-            out = v * evaluate(p)
+            out = evaluate(p)
+            if abs(v - 1.0) > eps:
+                out = v * out
             for v, p in vp_list[1:]:
-                out = out + v * evaluate(p)
+                t = evaluate(p)
+                if abs(v - 1.0) > eps:
+                    t = v * t
+                out = out + t
             outs.append(out)
 
         out = torch.cat(outs, dim=-1)